move code to product mutation
